Move Packing<string> into sources
I actually need the string packing implementation to do a set_union on a
`std::map<T, std::string>`. Moreover, we use string packing in several
tests so it seems of general use
